(useTextFilter) rename path properly
