+++ frontend/src/components/pages/search/hooks/useTextFilter.ts
