+++ frontend/src/components/pages/search/hooks/useTrekResults.ts
